[refactor] remover from service layer to common | Rules | Externals
