+++ src/app/Services/Transaction/TransactionServiceAsync.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\Transaction;
 
+use App\Common\ManageRule\ManageRulesInterface;
 use App\Domain\Financial\Transaction\Request\TransactionRequestInterface;
 use App\Domain\Financial\Transaction\Service\TransactionServiceAsyncInterface;
 use App\Jobs\Transaction\CreateTransaction;
@@ -13,12 +14,26 @@ use App\Jobs\Transaction\CreateTransaction;
 class TransactionServiceAsync implements TransactionServiceAsyncInterface
 {
     /**
+     * @var ManageRulesInterface
+     */
+    private ManageRulesInterface $manageRules;
+
+    /**
+     * TransactionServiceAsync constructor.
+     * @param ManageRulesInterface $manageRules
+     */
+    public function __construct(ManageRulesInterface $manageRules)
+    {
+        $this->manageRules = $manageRules;
+    }
+
+    /**
      * @param TransactionRequestInterface $request
      * @return bool
      */
     public function store(TransactionRequestInterface $request): array
     {
-        CreateTransaction::dispatchAfterResponse($request);
+        CreateTransaction::dispatchAfterResponse($request,$this->manageRules);
         return ['message' => 'started queues'];
 
     }
